+++ server/src/http/controllers/jobsEtFormations.controller.ts
@@ -1,5 +1,5 @@
 import { zRoutes } from "shared"
-import { LBA_ITEM_TYPE, allLbaItemType } from "shared/constants/lbaiten"
+import { LBA_ITEM_TYPE, allLbaItemType } from "shared/constants/lbaitem"
 
 import { trackApiCall } from "../../common/utils/sendTrackingEvent"
 import { getFormationsQuery } from "../../services/formation.service"
