+++ server/src/services/lbacompany.service.ts
@@ -1,6 +1,6 @@
 import Boom from "boom"
 import { ILbaCompany } from "shared"
-import { LBA_ITEM_TYPE } from "shared/constants/lbaiten"
+import { LBA_ITEM_TYPE } from "shared/constants/lbaitem"
 
 import { LbaCompany } from "../common/model/index"
 import { encryptMailWithIV } from "../common/utils/encryptString"
