+++ ui/components/ItemDetail/CandidatureLba/services/isCandidatureLba.ts
@@ -1,4 +1,4 @@
-import { LBA_ITEM_TYPE } from "shared/constants/lbaiten"
+import { LBA_ITEM_TYPE } from "shared/constants/lbaitem"
 
 import { isNonEmptyString } from "../../../../utils/strutils"
