+++ ui/components/ItemDetail/Training.tsx
@@ -1,6 +1,6 @@
 import { Box, Button, Flex, Image, Link, Text } from "@chakra-ui/react"
 import React, { useContext, useState } from "react"
-import { LBA_ITEM_TYPE } from "shared/constants/lbaiten"
+import { LBA_ITEM_TYPE } from "shared/constants/lbaitem"
 
 import { focusWithin } from "@/theme/theme-lba-tools"
