+++ ui/utils/itemListUtils.ts
@@ -1,6 +1,6 @@
 // retourne les offres issues du dépôt d'offre simplifié (ex Matcha) triées par ordre croissant de distance au centre de recherche
 
-import { LBA_ITEM_TYPE } from "shared/constants/lbaiten"
+import { LBA_ITEM_TYPE } from "shared/constants/lbaitem"
 
 // suivi des offres pe triées par ordre croissant de distance au centre de recherche.
 export const mergeJobs = ({ jobs, activeFilters }) => {
